refactor camera manager and unify implementation
