+++ tests/test_face_tracker.py
@@ -1,8 +1,9 @@
-import numpy as np
-import fakeredis
 import sys
 from pathlib import Path
 from types import SimpleNamespace
+
+import fakeredis
+import numpy as np
 
 ROOT = Path(__file__).resolve().parents[1]
 sys.path.insert(0, str(ROOT))
@@ -16,22 +17,25 @@ sys.modules.setdefault(
     ),
 )
 
-from modules.tracker.face_tracker import FaceTracker
-from modules import face_db
-from core import events
-from modules.camera_manager import CameraManager
-import asyncio
+import asyncio  # noqa: E402
+
+from core import events  # noqa: E402
+from core.camera_manager import CameraManager  # noqa: E402
+from modules import face_db  # noqa: E402
+from modules.tracker.face_tracker import FaceTracker  # noqa: E402
 
 
 def _make_detector(dets):
     class D:
         def __init__(self):
             self.dets = list(dets)
+
         def detect(self, frame):
             if not self.dets:
                 return []
             x1, y1, x2, y2, s = self.dets.pop(0)
             return [SimpleNamespace(bbox=(x1, y1, x2, y2), det_score=s)]
+
     return D()
 
 
@@ -39,11 +43,13 @@ def _make_model(out):
     class M:
         def __init__(self):
             self.out = list(out)
+
         def get(self, crop):
             if not self.out:
                 return []
             data = self.out.pop(0)
             return [SimpleNamespace(**data)]
+
     return M()
 
 
@@ -58,9 +64,7 @@ def test_face_tracker_flow(monkeypatch):
     monkeypatch.setattr(
         "modules.tracker.face_tracker._sharpness", lambda img: sharp.pop(0)
     )
-    monkeypatch.setattr(
-        "modules.tracker.face_tracker._frontalness", lambda face: 0.5
-    )
+    monkeypatch.setattr("modules.tracker.face_tracker._frontalness", lambda face: 0.5)
     det = _make_detector(
         [
             (20, 10, 50, 40, 0.8),
@@ -69,11 +73,13 @@ def test_face_tracker_flow(monkeypatch):
         ]
     )
     emb = np.ones(512, dtype=np.float32)
-    model = _make_model([
-        {"embedding": emb, "match_score": 0.5},
-        {"embedding": emb, "match_score": 0.5},
-        {"embedding": emb, "match_score": 0.5},
-    ])
+    model = _make_model(
+        [
+            {"embedding": emb, "match_score": 0.5},
+            {"embedding": emb, "match_score": 0.5},
+            {"embedding": emb, "match_score": 0.5},
+        ]
+    )
     ft = FaceTracker(1, {"cpu_sample_every": 1, "min_face_size": 10}, r)
     ft.detector = det
     ft.model = model
@@ -122,9 +128,9 @@ def test_camera_manager_starts_face_tracker():
             {},
             r,
             lambda: cams,
-            lambda cam, cfg, tr, r: tr.setdefault(cam["id"], object()),
+            lambda cam, cfg, tr, r, cb=None: tr.setdefault(cam["id"], object()),
             lambda cid, tr: tr.pop(cid, None),
-            lambda cam, cfg, tr, r: tr.setdefault(cam["id"], object()),
+            lambda cam, cfg, tr, r, cb=None: tr.setdefault(cam["id"], object()),
             lambda cid, tr: tr.pop(cid, None),
         )
         await mgr._start_tracker_background(cams[0])
